Created basic file upload service
Currently supports operations:
* Create file -> get metadata and ID (POST /files)
* Get file metadata by ID (GET /files/(id))
* Remove file and metadata by ID (DELETE /files/(id))

File data is stored in a directory on disk.
File metadata is stored in tables in SurrealDB.
The "id" field of the record ID is used as the file ID.
This saves us from needing an additional UUID field.

File upload size limit is 100 MB.

What remains to be done:
* Implement user auth
* Allow for setting upload path by environment variable